# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

/**
 * Primary technical interest area
 */
export type StudentInputTechnicalInterest = typeof StudentInputTechnicalInterest[keyof typeof StudentInputTechnicalInterest];


export const StudentInputTechnicalInterest = {
  web_development: 'web_development',
  data_analytics: 'data_analytics',
  ai_ml: 'ai_ml',
  cyber_security: 'cyber_security',
  cloud_computing: 'cloud_computing',
  app_development: 'app_development',
  blockchain: 'blockchain',
  devops: 'devops',
  game_development: 'game_development',
  ui_ux_design: 'ui_ux_design',
  iot: 'iot',
  ar_vr: 'ar_vr',
} as const;

export interface StudentInput {
  /** Student's name */
  name: string;
  /**
     * Attendance percentage
     * @minimum 0
     * @maximum 100
     */
  attendance: number;
  /**
     * Coding skill score (0-100)
     * @minimum 0
     * @maximum 100
     */
  codingSkill: number;
  /**
     * Aptitude test score (0-100)
     * @minimum 0
     * @maximum 100
     */
  aptitudeScore: number;
  /**
     * Communication skill score (0-100)
     * @minimum 0
     * @maximum 100
     */
  communicationSkill: number;
  /**
     * Overall academic/CGPA score (0-100)
     * @minimum 0
     * @maximum 100
     */
  academicScore: number;
  /** Primary technical interest area */
  technicalInterest: StudentInputTechnicalInterest;
  /** Student's GitHub profile URL */
  githubUrl?: string | null;
  /** Student's LinkedIn profile URL */
  linkedinUrl?: string | null;
}

/**
 * Student skill level classification
 */
export type CareerAssessmentSkillGroup = typeof CareerAssessmentSkillGroup[keyof typeof CareerAssessmentSkillGroup];


export const CareerAssessmentSkillGroup = {
  Beginner: 'Beginner',
  Intermediate: 'Intermediate',
  Advanced: 'Advanced',
} as const;

export interface CareerAssessment {
  id: number;
  studentName: string;
  attendance: number;
  codingSkill: number;
  aptitudeScore: number;
  communicationSkill: number;
  academicScore: number;
  technicalInterest: string;
  githubUrl?: string | null;
  linkedinUrl?: string | null;
  /** Predicted overall performance score (0-100) */
  performanceScore: number;
  /** Whether the student is eligible for placement */
  placementEligible: boolean;
  /** Recommended career domain */
  recommendedDomain: string;
  /** Student skill level classification */
  skillGroup: CareerAssessmentSkillGroup;
  /** AI-generated personalized career improvement suggestions */
  aiSuggestions: string[];
  createdAt: string;
}

export interface AssessmentSummary {
  id: number;
  studentName: string;
  performanceScore: number;
  placementEligible: boolean;
  recommendedDomain: string;
  skillGroup: string;
  createdAt: string;
}

export interface ScoreHistoryPoint {
  date: string;
  score: number;
  domain: string;
  skillGroup: string;
}

export interface UserProfile {
  userId: string;
  displayName: string;
  totalAssessments: number;
  avgScore: number;
  bestScore: number;
  placementRate: number;
  scoreHistory: ScoreHistoryPoint[];
  assessments: AssessmentSummary[];
}

export type CareerStatsSkillGroupCounts = {
  Beginner: number;
  Intermediate: number;
  Advanced: number;
};

export type CareerStatsDomainDistributionItem = {
  domain: string;
  count: number;
};

export interface CareerStats {
  totalAssessments: number;
  placementEligibleCount: number;
  /** Percentage of students eligible for placement */
  placementEligibleRate: number;
  averagePerformanceScore: number;
  skillGroupCounts: CareerStatsSkillGroupCounts;
  domainDistribution: CareerStatsDomainDistributionItem[];
  recentAssessments: AssessmentSummary[];
}

export interface CareerDomain {
  id: string;
  name: string;
  description: string;
  keySkills: string[];
  avgSalaryRange: string;
}

export type AuthUserUser = {
  id: string;
  username: string;
  firstName?: string | null;
  lastName?: string | null;
  profileImage?: string | null;
} | null;

export interface AuthUser {
  isAuthenticated: boolean;
  user?: AuthUserUser;
}

export interface ApiError {
  error: string;
}