# lib/api-zod/src/generated/types/authUser.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUserUser } from './authUserUser';

export interface AuthUser {
  isAuthenticated: boolean;
  user?: AuthUserUser;
}