# lib/api-zod/src/generated/types/careerAssessment.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */
import type { CareerAssessmentSkillGroup } from './careerAssessmentSkillGroup';

export interface CareerAssessment {
  id: number;
  studentName: string;
  attendance: number;
  codingSkill: number;
  aptitudeScore: number;
  communicationSkill: number;
  academicScore: number;
  technicalInterest: string;
  githubUrl?: string | null;
  linkedinUrl?: string | null;
  /** Predicted overall performance score (0-100) */
  performanceScore: number;
  /** Whether the student is eligible for placement */
  placementEligible: boolean;
  /** Recommended career domain */
  recommendedDomain: string;
  /** Student skill level classification */
  skillGroup: CareerAssessmentSkillGroup;
  /** AI-generated personalized career improvement suggestions */
  aiSuggestions: string[];
  createdAt: Date;
}