# lib/api-zod/src/generated/types/careerDomain.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */

export interface CareerDomain {
  id: string;
  name: string;
  description: string;
  keySkills: string[];
  avgSalaryRange: string;
}