# lib/api-zod/src/generated/types/scoreHistoryPoint.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreHistoryPoint {
  date: Date;
  score: number;
  domain: string;
  skillGroup: string;
}