# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */
import type { StudentInputTechnicalInterest } from './studentInputTechnicalInterest';

export interface StudentInput {
  /** Student's name */
  name: string;
  /**
     * Attendance percentage
     * @minimum 0
     * @maximum 100
     */
  attendance: number;
  /**
     * Coding skill score (0-100)
     * @minimum 0
     * @maximum 100
     */
  codingSkill: number;
  /**
     * Aptitude test score (0-100)
     * @minimum 0
     * @maximum 100
     */
  aptitudeScore: number;
  /**
     * Communication skill score (0-100)
     * @minimum 0
     * @maximum 100
     */
  communicationSkill: number;
  /**
     * Overall academic/CGPA score (0-100)
     * @minimum 0
     * @maximum 100
     */
  academicScore: number;
  /** Primary technical interest area */
  technicalInterest: StudentInputTechnicalInterest;
  /** Student's GitHub profile URL */
  githubUrl?: string | null;
  /** Student's LinkedIn profile URL */
  linkedinUrl?: string | null;
}