# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AI Career Guidance System API
 * OpenAPI spec version: 0.1.0
 */
import type { AssessmentSummary } from './assessmentSummary';
import type { ScoreHistoryPoint } from './scoreHistoryPoint';

export interface UserProfile {
  userId: string;
  displayName: string;
  totalAssessments: number;
  avgScore: number;
  bestScore: number;
  placementRate: number;
  scoreHistory: ScoreHistoryPoint[];
  assessments: AssessmentSummary[];
}